mv fileinfo.CopyFile to fsx.
